Label objects with unique id
adding a unique id to the label name enables users to uniquely identify the printed instance. this is for example helpful when a user wishes to exclude an object during print. previously, selecting an object to exclude may select multiple objects of the same model when the instances of the object are in two different copy of the same model